Moved FileInputFormat and MutipleInputs to ao.adoop.mapreduce
